fix: shim fetch only for integration tests
Shimming it for both unit and integration
was having almost 80% increase in execution
time for unit tests.
